Allow times in created_at dates
There are a couple of places where I've written more than one article on
the same day.  Now I can add times to the `created_at` value for those
articles to force a particular sort order in the archive list.  I had to
update the code that outputs dates to ensure that they're still printed
in a consistent way.  (In particular, I don't want to display the time,
since they're not really accurate; they're just there to enforce an
ordering.)